WIP
Mostly left comments directly inline, was easiest.

---

When installing the requirements, I'm getting the following error in the output:
```
ERROR: pip's dependency resolver does not currently take into account all the packages that are installed. This behaviour is the source of the following dependency conflicts.
frictionless 5.16.0 requires chardet>=3.0, which is not installed.
frictionless 5.16.0 requires jsonschema<4.18, which is not installed.
frictionless 5.16.0 requires requests>=2.10, which is not installed.
visidata 3.0 requires importlib-metadata>=3.6, which is not installed.
frictionless 5.16.0 requires pydantic>=2.0, but you have pydantic 1.9.0 which is incompatible.
```

However, the exit code is successful.

---

...